requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
black==22.1.0
click==8.0.3
colorama==0.4.4
crispy-bootstrap5==0.6
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.2.12
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-js-asset==2.0.0
flake8==4.0.1
gunicorn==20.1.0
isort==5.10.1
lxml==4.7.1
markdown2==2.4.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pyquery==1.4.3
python-dotenv==0.19.2
pytz==2021.3
readtime==1.1.1
soupsieve==2.3.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.0.1
whitenoise==6.0.0
